+++ src/SHADERed/UI/Tools/DebuggerOutline.cpp
@@ -25,14 +25,14 @@ namespace ed {
 		}
 
 		for (i = 0, j = 1; j < pixel.VertexCount; ++i, ++j)
-			drawList->AddLine(ImVec2(uiPos.x + vPos[i].x, uiPos.y + vPos[i].y), ImVec2(uiPos.x + vPos[j].x, uiPos.y + vPos[j].y), outlineColor);
+			drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(vPos[i].x), uiPos.y + static_cast<float>(vPos[i].y)), ImVec2(uiPos.x + static_cast<float>(vPos[j].x), uiPos.y + static_cast<float>(vPos[j].y)), outlineColor);
 		if (pixel.VertexCount > 2)
-			drawList->AddLine(ImVec2(uiPos.x + vPos[i].x, uiPos.y + vPos[i].y), ImVec2(uiPos.x + vPos[0].x, uiPos.y + vPos[0].y), outlineColor);
+			drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(vPos[i].x), uiPos.y + static_cast<float>(vPos[i].y)), ImVec2(uiPos.x + static_cast<float>(vPos[0].x), uiPos.y + static_cast<float>(vPos[0].y)), outlineColor);
 
 		for (i = 0; i < pixel.VertexCount; ++i) {
 			char labelChar = static_cast<char>('0' + i);
 			std::string label(1, labelChar);
-			drawList->AddText(ImVec2(uiPos.x + vPos[i].x, uiPos.y + vPos[i].y), outlineColor, label.c_str());
+			drawList->AddText(ImVec2(uiPos.x + static_cast<float>(vPos[i].x), uiPos.y + static_cast<float>(vPos[i].y)), outlineColor, label.c_str());
 		}
 
 		if (pixel.GeometryShaderUsed) {
@@ -54,12 +54,12 @@ namespace ed {
 			for (i = 0; i < gsOutputsCount; ++i)
 				gsPos[i] = (getScreenCoordInverted(pixel.FinalPosition[i]) - glm::vec2(zoomPos.x, -zoomPos.y)) * (1.0f / zoomSize) * itemSize;
 			for (i = 0, j = 1; j < gsOutputsCount; ++i, ++j)
-				drawList->AddLine(ImVec2(uiPos.x + gsPos[i].x, uiPos.y + gsPos[i].y), ImVec2(uiPos.x + gsPos[j].x, uiPos.y + gsPos[j].y), outlineColor);
+				drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(gsPos[i].x), uiPos.y + static_cast<float>(gsPos[i].y)), ImVec2(uiPos.x + static_cast<float>(gsPos[j].x), uiPos.y + static_cast<float>(gsPos[j].y)), outlineColor);
 			if (gsOutputsCount > 2)
-				drawList->AddLine(ImVec2(uiPos.x + gsPos[i].x, uiPos.y + gsPos[i].y), ImVec2(uiPos.x + gsPos[0].x, uiPos.y + gsPos[0].y), outlineColor);
+				drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(gsPos[i].x), uiPos.y + static_cast<float>(gsPos[i].y)), ImVec2(uiPos.x + static_cast<float>(gsPos[0].x), uiPos.y + static_cast<float>(gsPos[0].y)), outlineColor);
 			for (i = 0; i < gsOutputsCount; ++i) {
 				const char GS0[4] = { 'G', 'S', static_cast<char>('0' + i), '\0' };
-				drawList->AddText(ImVec2(uiPos.x + gsPos[i].x, uiPos.y + gsPos[i].y), outlineColor, GS0);
+				drawList->AddText(ImVec2(uiPos.x + static_cast<float>(gsPos[i].x), uiPos.y + static_cast<float>(gsPos[i].y)), outlineColor, GS0);
 			}
 		}
 	}
@@ -77,21 +77,21 @@ namespace ed {
 		float hfpxSizeY = 0.5f / zoomSize.y;
 
 		// lines
-		drawList->AddLine(ImVec2(uiPos.x, uiPos.y + pxPos.y), ImVec2(uiPos.x + pxPos.x - hfpxSizeY, uiPos.y + pxPos.y), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x + hfpxSizeY, uiPos.y + pxPos.y), ImVec2(uiPos.x + itemSize.x, uiPos.y + pxPos.y), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x, uiPos.y), ImVec2(uiPos.x + pxPos.x, uiPos.y + pxPos.y - hfpxSizeY), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x, uiPos.y + pxPos.y + hfpxSizeY), ImVec2(uiPos.x + pxPos.x, uiPos.y + itemSize.y), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x, uiPos.y + static_cast<float>(pxPos.y)), ImVec2(uiPos.x + static_cast<float>(pxPos.x) - hfpxSizeY, uiPos.y + static_cast<float>(pxPos.y)), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeY, uiPos.y + static_cast<float>(pxPos.y)), ImVec2(uiPos.x + itemSize.x, uiPos.y + static_cast<float>(pxPos.y)), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x), uiPos.y), ImVec2(uiPos.x + static_cast<float>(pxPos.x), uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x), uiPos.y + static_cast<float>(pxPos.y) + hfpxSizeY), ImVec2(uiPos.x + static_cast<float>(pxPos.x), uiPos.y + itemSize.y), outlineColor);
 
 		// rectangle
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x - hfpxSizeX, uiPos.y + pxPos.y - hfpxSizeY), ImVec2(uiPos.x + pxPos.x - hfpxSizeX, uiPos.y + pxPos.y + hfpxSizeY), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x + hfpxSizeX, uiPos.y + pxPos.y - hfpxSizeY), ImVec2(uiPos.x + pxPos.x + hfpxSizeX, uiPos.y + pxPos.y + hfpxSizeY), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x - hfpxSizeX, uiPos.y + pxPos.y + hfpxSizeY), ImVec2(uiPos.x + pxPos.x + hfpxSizeX, uiPos.y + pxPos.y + hfpxSizeY), outlineColor);
-		drawList->AddLine(ImVec2(uiPos.x + pxPos.x - hfpxSizeX, uiPos.y + pxPos.y - hfpxSizeY), ImVec2(uiPos.x + pxPos.x + hfpxSizeX, uiPos.y + pxPos.y - hfpxSizeY), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x) - hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), ImVec2(uiPos.x + static_cast<float>(pxPos.x) - hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) + hfpxSizeY), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) + hfpxSizeY), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x) - hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) + hfpxSizeY), ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) + hfpxSizeY), outlineColor);
+		drawList->AddLine(ImVec2(uiPos.x + static_cast<float>(pxPos.x) - hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), outlineColor);
 
 		if (zoomSize.x < 0.3f) {
 			static char pxCoord[32] = { 0 };
 			snprintf(pxCoord, sizeof(pxCoord), "(%d, %d)", pixel.Coordinate.x, pixel.Coordinate.y);
-			drawList->AddText(ImVec2(uiPos.x + pxPos.x + hfpxSizeX, uiPos.y + pxPos.y - hfpxSizeY), outlineColor, pxCoord);
+			drawList->AddText(ImVec2(uiPos.x + static_cast<float>(pxPos.x) + hfpxSizeX, uiPos.y + static_cast<float>(pxPos.y) - hfpxSizeY), outlineColor, pxCoord);
 		}
 	}
 }
